# src/main/resources/swagger/api.json
{
  "swagger": "2.0",
  "info": {
    "version": "0.0.1-SNAPSHOT",
    "title": "atlas",
    "x-appId": "1000001128",
    "x-serviceName": "atlas"
  },
  "host": "localhost:8081",
  "basePath": "/",
...
  "paths": {
...
    "/web/apply/createApply": {
      "post": {
        "tags": [
          "apply-controller"
        ],
        "summary": "createApply",
        "operationId": "createApplyUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "applyDto",
            "description": "applyDto",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ApplyDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Response«ApplyDto»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/web/apply/queryByPage": {
      "get": {
        "tags": [
          "apply-controller"
        ],
        "summary": "queryByPage",
        "operationId": "queryByPageUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "applyUser",
            "in": "query",
            "description": "applyUser",
            "required": false,
            "type": "string"
          },
          {
            "name": "status",
            "in": "query",
            "description": "status",
            "required": false,
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "description": "page",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "size",
            "in": "query",
            "description": "size",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Response«PageVO«ApplyDto»»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/web/apply/updateStatus": {
      "post": {
        "tags": [
          "apply-controller"
        ],
        "summary": "updateStatus",
        "operationId": "updateStatusUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "request",
            "description": "request",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateApplyStatusDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Response«string»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/web/appquotas": {
      "post": {
        "tags": [
          "app-quota-controller"
        ],
        "summary": "addQuota",
        "operationId": "addQuotaUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "appQuotaDto",
            "description": "appQuotaDto",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AppQuotaDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Response«string»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "app-quota-controller"
        ],
        "summary": "updateQuota",
        "operationId": "updateQuotaUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "appQuotaDto",
            "description": "appQuotaDto",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AppQuotaDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Response«string»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/web/appquotas/all": {
      "get": {
        "tags": [
          "app-quota-controller"
        ],
        "summary": "getAllQuotas",
        "operationId": "getAllQuotasUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Response«List«AppQuotaDto»»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/web/appquotas/page": {
      "get": {
        "tags": [
          "app-quota-controller"
        ],
        "summary": "getQuotasByPage",
        "operationId": "getQuotasByPageUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "page",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "size",
            "in": "query",
            "description": "size",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Response«PageDto«AppQuotaDto»»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/web/quotas": {
      "post": {
        "tags": [
          "quota-controller"
        ],
        "summary": "addQuota",
        "operationId": "addQuotaUsingPOST_1",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "quotaDtoPlus",
            "description": "quotaDtoPlus",
            "required": true,
            "schema": {
              "$ref": "#/definitions/QuotaDtoPlus"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Response«QuotaDtoPlus»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "quota-controller"
        ],
        "summary": "updateQuota",
        "operationId": "updateQuotaUsingPUT_1",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "quotaDtoPlus",
            "description": "quotaDtoPlus",
            "required": true,
            "schema": {
              "$ref": "#/definitions/QuotaDtoPlus"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Response«QuotaDtoPlus»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/web/quotas/all": {
      "get": {
        "tags": [
          "quota-controller"
        ],
        "summary": "getAllQuotas",
        "operationId": "getAllQuotasUsingGET_1",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Response«List«QuotaDtoPlus»»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/web/quotas/condition": {
      "get": {
        "tags": [
          "quota-controller"
        ],
        "summary": "getQuotasByCondition",
        "operationId": "getQuotasByConditionUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "page",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "size",
            "in": "query",
            "description": "size",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "orgName",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "envId",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Response«PageDto«QuotaDtoPlus»»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/web/quotas/page": {
      "get": {
        "tags": [
          "quota-controller"
        ],
        "summary": "getQuotasByPage",
        "operationId": "getQuotasByPageUsingGET_1",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "page",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "size",
            "in": "query",
            "description": "size",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Response«PageDto«QuotaDtoPlus»»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/web/quotas/queryquota": {
      "get": {
        "tags": [
          "quota-controller"
        ],
        "summary": "getQuotaByOrgEnvSpec",
        "operationId": "getQuotaByOrgEnvSpecUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "orgCode",
            "in": "query",
            "description": "orgCode",
            "required": true,
            "type": "string"
          },
          {
            "name": "envName",
            "in": "query",
            "description": "envName",
            "required": true,
            "type": "string"
          },
          {
            "name": "specName",
            "in": "query",
            "description": "specName",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Response«QuotaDtoPlus»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/web/quotas/quota": {
      "get": {
        "tags": [
          "quota-controller"
        ],
        "summary": "getQuotaByEnvAndOrg",
        "operationId": "getQuotaByEnvAndOrgUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "orgCode",
            "in": "query",
            "description": "orgCode",
            "required": true,
            "type": "string"
          },
          {
            "name": "envName",
            "in": "query",
            "description": "envName",
            "required": true,
            "type": "string"
          },
          {
            "name": "specName",
            "in": "query",
            "description": "specName",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Response«QuotaDto»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/web/quotas/{quotaId}": {
      "get": {
        "tags": [
          "quota-controller"
        ],
        "summary": "getQuotaById",
        "operationId": "getQuotaByIdUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "quotaId",
            "in": "path",
            "description": "quotaId",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Response«QuotaDtoPlus»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "quota-controller"
        ],
        "summary": "deleteQuota",
        "operationId": "deleteQuotaUsingDELETE",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "quotaId",
            "in": "path",
            "description": "quotaId",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Response«string»"
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "AppQuotaDto": {
      "type": "object",
      "properties": {
        "active": {
          "type": "boolean"
        },
        "appId": {
          "type": "string"
        },
        "appName": {
          "type": "string"
        },
        "envId": {
          "type": "integer",
          "format": "int64"
        },
        "envName": {
          "type": "string"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "insertBy": {
          "type": "string"
        },
        "insertTime": {
          "type": "string",
          "format": "date-time"
        },
        "number": {
          "type": "integer",
          "format": "int64"
        },
        "orgId": {
          "type": "integer",
          "format": "int64"
        },
        "orgName": {
          "type": "string"
        },
        "spectypeId": {
          "type": "integer",
          "format": "int64"
        },
        "spectypeName": {
          "type": "string"
        },
        "updateBy": {
          "type": "string"
        },
        "updateTime": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "ApplyDto": {
      "type": "object",
      "properties": {
        "applyDepartment": {
          "type": "string"
        },
        "applyTime": {
          "type": "integer",
          "format": "int64"
        },
        "applyUser": {
          "type": "string"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "opUser": {
          "type": "string"
        },
        "request": {
          "type": "string"
        },
        "result": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "type": {
          "type": "string"
        }
      }
    },
    "EnvDto": {
      "type": "object",
      "properties": {
        "active": {
          "type": "boolean"
        },
        "consul": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "dns": {
          "type": "string"
        },
        "dockeryard": {
          "type": "string"
        },
        "extensions": {
          "type": "object"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "insertBy": {
          "type": "string"
        },
        "insertTime": {
          "type": "string",
          "format": "date-time"
        },
        "name": {
          "type": "string"
        },
        "nginx": {
          "type": "string"
        },
        "updateBy": {
          "type": "string"
        },
        "updateTime": {
          "type": "string",
          "format": "date-time"
        },
        "zones": {
          "type": "string"
        }
      }
    },
...
    "OrgDto": {
      "type": "object",
      "properties": {
        "active": {
          "type": "boolean"
        },
        "description": {
          "type": "string"
        },
        "extensions": {
          "type": "object"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "insertBy": {
          "type": "string"
        },
        "insertTime": {
          "type": "string",
          "format": "date-time"
        },
        "name": {
          "type": "string"
        },
        "orgCode": {
          "type": "string"
        },
        "parentOrgId": {
          "type": "integer",
          "format": "int64"
        },
        "updateBy": {
          "type": "string"
        },
        "updateTime": {
          "type": "string",
          "format": "date-time"
        },
        "userDto": {
          "$ref": "#/definitions/UserDto"
        },
        "userWorkNumber": {
          "type": "string"
        }
      }
    },
...
    "PageDto«AppQuotaDto»": {
      "type": "object",
      "properties": {
        "content": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/AppQuotaDto"
          }
        },
        "first": {
          "type": "boolean"
        },
        "last": {
          "type": "boolean"
        },
        "number": {
          "type": "integer",
          "format": "int32"
        },
        "numberOfElements": {
          "type": "integer",
          "format": "int32"
        },
        "size": {
          "type": "integer",
          "format": "int32"
        },
        "totalElements": {
          "type": "integer",
          "format": "int64"
        },
        "totalPages": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
...
    "PageDto«QuotaDtoPlus»": {
      "type": "object",
      "properties": {
        "content": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/QuotaDtoPlus"
          }
        },
        "first": {
          "type": "boolean"
        },
        "last": {
          "type": "boolean"
        },
        "number": {
          "type": "integer",
          "format": "int32"
        },
        "numberOfElements": {
          "type": "integer",
          "format": "int32"
        },
        "size": {
          "type": "integer",
          "format": "int32"
        },
        "totalElements": {
          "type": "integer",
          "format": "int64"
        },
        "totalPages": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
...
    "PageVO«ApplyDto»": {
      "type": "object",
      "properties": {
        "content": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ApplyDto"
          }
        },
        "first": {
          "type": "boolean"
        },
        "last": {
          "type": "boolean"
        },
        "number": {
          "type": "integer",
          "format": "int32"
        },
        "numberOfElements": {
          "type": "integer",
          "format": "int32"
        },
        "size": {
          "type": "integer",
          "format": "int32"
        },
        "totalElements": {
          "type": "integer",
          "format": "int64"
        },
        "totalPages": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "QuotaDto": {
      "type": "object",
      "properties": {
        "active": {
          "type": "boolean"
        },
        "balance": {
          "type": "number",
          "format": "float"
        },
        "description": {
          "type": "string"
        },
        "envDto": {
          "$ref": "#/definitions/EnvDto"
        },
        "envId": {
          "type": "integer",
          "format": "int64"
        },
        "extensions": {
          "type": "object"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "insertBy": {
          "type": "string"
        },
        "insertTime": {
          "type": "string",
          "format": "date-time"
        },
        "limit": {
          "type": "number",
          "format": "float"
        },
        "orgDto": {
          "$ref": "#/definitions/OrgDto"
        },
        "orgId": {
          "type": "integer",
          "format": "int64"
        },
        "orgName": {
          "type": "string"
        },
        "specTypeDto": {
          "$ref": "#/definitions/SpecTypeDto"
        },
        "specTypeId": {
          "type": "integer",
          "format": "int64"
        },
        "updateBy": {
          "type": "string"
        },
        "updateTime": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "QuotaDtoPlus": {
      "type": "object",
      "properties": {
        "active": {
          "type": "boolean"
        },
        "cpu": {
          "type": "integer",
          "format": "int64"
        },
        "description": {
          "type": "string"
        },
        "disk": {
          "type": "integer",
          "format": "int64"
        },
        "envDto": {
          "$ref": "#/definitions/EnvDto"
        },
        "envId": {
          "type": "integer",
          "format": "int64"
        },
        "extensions": {
          "type": "object"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "insertBy": {
          "type": "string"
        },
        "insertTime": {
          "type": "string",
          "format": "date-time"
        },
        "memory": {
          "type": "integer",
          "format": "int64"
        },
        "orgDto": {
          "$ref": "#/definitions/OrgDto"
        },
        "orgId": {
          "type": "integer",
          "format": "int64"
        },
        "orgName": {
          "type": "string"
        },
        "updateBy": {
          "type": "string"
        },
        "updateTime": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
...
    "Response«ApplyDto»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "details": {
          "$ref": "#/definitions/ApplyDto"
        },
        "message": {
          "type": "string"
        }
      }
    },
...
    "Response«List«AppQuotaDto»»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "detail": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/AppQuotaDto"
          }
        },
        "message": {
          "type": "string"
        }
      }
    },
...
    "Response«List«QuotaDtoPlus»»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "detail": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/QuotaDtoPlus"
          }
        },
        "message": {
          "type": "string"
        }
      }
    },
...
    "Response«PageDto«AppQuotaDto»»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "detail": {
          "$ref": "#/definitions/PageDto«AppQuotaDto»"
        },
        "message": {
          "type": "string"
        }
      }
    },
...
    "Response«PageDto«QuotaDtoPlus»»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "detail": {
          "$ref": "#/definitions/PageDto«QuotaDtoPlus»"
        },
        "message": {
          "type": "string"
        }
      }
    },
...
    "Response«PageVO«ApplyDto»»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "details": {
          "$ref": "#/definitions/PageVO«ApplyDto»"
        },
        "message": {
          "type": "string"
        }
      }
    },
    "Response«QuotaDtoPlus»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "detail": {
          "$ref": "#/definitions/QuotaDtoPlus"
        },
        "message": {
          "type": "string"
        }
      }
    },
    "Response«QuotaDto»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "detail": {
          "$ref": "#/definitions/QuotaDto"
        },
        "message": {
          "type": "string"
        }
      }
    },
...
    "Response«string»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "detail": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    },
...
    "UpdateApplyStatusDto": {
      "type": "object",
      "properties": {
        "applyId": {
          "type": "integer",
          "format": "int64"
        },
        "opUser": {
          "type": "string"
        },
        "result": {
          "type": "string"
        },
        "status": {
          "type": "string"
        }
      }
    },
...
  }
}